verify board existence

+++ src/modules/board/service/implementation/BoardService.ts
@@ -4,6 +4,7 @@ import { IBoardUpdateDTO } from "@modules/board/dtos/IBoardUpdateDTO";
 import { IBoardService } from "../interface/IBoardService";
 import { inject, injectable } from "tsyringe";
 import { IBoardRepository } from "@modules/board/infra/database/repository/IBoardRepository";
+import AppError from "@shared/infra/error/AppError";
 
 @injectable()
 class BoardService implements IBoardService {
@@ -18,10 +19,18 @@ class BoardService implements IBoardService {
     }
     
     public async update(boardDTO: IBoardUpdateDTO): Promise<IBoardResponseDTO | undefined> {
+        const board = await this.boardRepository.findById(boardDTO.id);
+
+        if (!board) throw new AppError('Board not found!', 404);
+
         return await this.boardRepository.update(boardDTO);
     }
 
     public async delete(id: number): Promise<boolean> {
+        const board = await this.boardRepository.findById(id);
+
+        if (!board) throw new AppError('Board not found!', 404);
+
         return await this.boardRepository.delete(id);
     }
